pyproject.toml
[project]
name = "langchain-project"
version = "0.1.0"
description = "LangChain project with OpenAI and Google GenAI integration"
authors = [{name = "Alcir Junior"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "pypdf>=5.0.0",
    "langchain-postgres>=0.0.16",
    "rich>=14.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.7.0",
    "mypy>=1.8.0",
]

[project.scripts]
langchain = "langchain_project.main:main"
langchain-project = "langchain_project.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/langchain_project"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
